test unit localbucketindex

+++ puma/puma-server/src/test/java/com/dianping/puma/storage/LocalBucketIndexTest.java
@@ -33,20 +33,22 @@ public class LocalBucketIndexTest {
 
 		work = new File(System.getProperty("java.io.tmpdir", "."), "Puma");
 
-		for (File file : work.listFiles()) {
-			for (File files : file.listFiles()) {
-				if (files.isFile()) {
-					if (files.delete()) {
-						System.out.println("Clear file: " + file.getAbsolutePath());
+		if (work != null && work.isDirectory()) {
+			for (File file : work.listFiles()) {
+				for (File files : file.listFiles()) {
+					if (files.isFile()) {
+						if (files.delete()) {
+							System.out.println("Clear file: " + file.getAbsolutePath());
+						}
 					}
 				}
+				if (file.delete()) {
+					System.out.println("Clear file: " + file.getAbsolutePath());
+				}
 			}
-			if (file.delete()) {
-				System.out.println("Clear file: " + file.getAbsolutePath());
+			if (work.delete()) {
+				System.out.println("Clear file: " + work.getAbsolutePath());
 			}
-		}
-		if (work.delete()) {
-			System.out.println("Clear file: " + work.getAbsolutePath());
 		}
 
 		work = new File(System.getProperty("java.io.tmpdir", "."), "Puma/20120710/bucket-0");
@@ -65,7 +67,7 @@ public class LocalBucketIndexTest {
 			e1.printStackTrace();
 		}
 
-		localBucketIndex.setBaseDir(System.getProperty("java.io.tmpdir", ".") + "Puma");
+		localBucketIndex.setBaseDir(System.getProperty("java.io.tmpdir", ".") + "/Puma");
 		localBucketIndex.setBucketFilePrefix("bucket-");
 		localBucketIndex.setMaxBucketLengthMB(500);
 
@@ -482,7 +484,7 @@ public class LocalBucketIndexTest {
 		System.out.println("***********************testGetBaseDir*************************");
 		System.out.println("**************************************************************");
 		this.localBucketIndex.init();
-		Assert.assertEquals((System.getProperty("java.io.tmpdir", ".").toString() + "Puma"), this.localBucketIndex
+		Assert.assertEquals((System.getProperty("java.io.tmpdir", ".").toString() + "/Puma"), this.localBucketIndex
 				.getBaseDir());
 		System.out.println("*************************************************************");
 		System.out.println("****************************End******************************");
